+++ R/plot_mid_importance.R
@@ -35,7 +35,7 @@ plot.mid.importance <- function(
     imp <- imp[1L:min(max.bars, nrow(imp), na.rm = TRUE), ]
     cols <- if (use.theme) {
       if (theme$type == "qualitative")
-        to.colors(imp$degree, theme)
+        to.colors(imp$order, theme)
       else
         to.colors(imp$importance, theme)
     } else "gray35"