+++ R/shapviz_mid.R
@@ -5,7 +5,7 @@
 #' The S3 method of \code{shapviz()} for the "mid" objects returns an object of class "shapviz" to be used to create SHAP plots with the functions of the shapviz package such as \code{sv_waterfall()} and \code{sv_importance()}.
 #'
 #' @param object a "mid" object.
-#' @param data a data frame containing observations for which SHAP values are calculated.
+#' @param data a data frame containing observations for which SHAP values are calculated. If not passed, data is extracted from \code{parent.env()} based on the function call of the "mid" object.
 #' @returns
 #' \code{shapviz.mid()} returns an object of class "shapviz".
 #' @exportS3Method shapviz::shapviz
@@ -25,8 +25,10 @@ shapviz.mid <- function(object, data = NULL) {
       shaps[, tags[2L]] <- shaps[, tags[2L]] + preds[, term] / 2
     }
   }
+  if (is.null(data))
+    data <- model.data(object)
   if (is.null(data)) {
-    message("'data' is not passed")
+    message("'data' not passed")
     data <- matrix(nrow = nrow(shaps), ncol = ncol(shaps))
     colnames(data) <- xvars
     data <- as.data.frame(data)