+++ tests/testthat/test-interpret.R
@@ -4,7 +4,7 @@ test_that("interpret() returns weighted least-norm solutions", {
   x2 = 1L:2L
   mid <- interpret(formula = x1^2 ~ x1 + x2,
                    singular.ok = TRUE)
-  expect_equal(mid$uninterpreted.rate, 0L)
+  expect_equal(mid$uninterpreted.variation, 0L)
   expect_equal(mid$main.effects$x1$mid, c(-.75, .75),
                ignore_attr = TRUE)
   # test 1 : rank deficient case 2
@@ -12,7 +12,7 @@ test_that("interpret() returns weighted least-norm solutions", {
   x2 <- c(1, -1, 2, -2, 1, -1, 2, -2)
   mid <- interpret(formula = x1^2 ~ x1 + x2,
                    singular.ok = TRUE)
-  expect_equal(mid$uninterpreted.rate, 0L)
+  expect_equal(mid$uninterpreted.variation, 0L)
   expect_equal(mid$main.effects$x1$mid, c(.75, -.75, -.75, .75),
                ignore_attr = TRUE)
   # test 3 : rank deficient case 3
@@ -21,7 +21,7 @@ test_that("interpret() returns weighted least-norm solutions", {
   x3 <- c(1, 2, 2, 2)
   mid <- interpret(formula = x1 * x2 * x3 ~ x1 + x2 + x3,
                    singular.ok = TRUE)
-  expect_equal(mid$uninterpreted.rate, 0L)
+  expect_equal(mid$uninterpreted.variation, 0L)
   expect_equal(mid$main.effects$x1$mid, c(-.45, .45),
                ignore_attr = TRUE)
 })
@@ -60,13 +60,11 @@ test_that("interpret() runs successfully for factor responses", {
   x <- 1:5
   y <- factor(c("a", "a", "a", "b", "b"), levels = c("b", "a"))
   mid <- interpret(y ~ x)
-  expect_equal(mid$intercept, 0.4)
-  expect_equal(mid$target.level, "b")
+  expect_equal(mid$intercept, 0.6)
   # test 2
   f <- function(X.model, newdata) ifelse(newdata$x < 4, "alice", "bob")
   mid <- interpret(object = NULL, x = x, pred.fun = f)
-  expect_equal(mid$intercept, 0.6)
-  expect_equal(mid$target.level, "alice")
+  expect_equal(mid$intercept, 0.4)
 })